requirements.txt
absl-py==1.3.0
addict==2.4.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
asgiref==3.5.2
basicsr==1.4.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
crcmod==1.7
cryptography==38.0.3
Django==4.1.3
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
imageio==2.22.4
jmespath==0.10.0
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
oss2==2.16.0
packaging==21.3
Pillow==9.3.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
tb-nightly==2.12.0a20221117
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.10.10
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
yapf==0.32.0
